Create README.md
Update README with project description and citation information

This commit adds a more detailed description of the project and its goals, including information on the data used and the models trained. It also includes citation information for the paper describing the work, as well as instructions for how to use the code in the repository.